Page List: Remove duplicate code
